+++ CHANGES.rst
@@ -4,6 +4,8 @@ History of changes to archive-tools
 dev (not yet released)
     New features
       + #35: class FileInfo calculates checksums lazily.
+      + #34: files of unsupported type are ignored when creating an
+	archive.  A warning is emitted instead of raising an error.
 
 0.2 (2019-07-14)
     New features